+++ demo-app/resources/js/Pages/Visit.vue
@@ -1,6 +1,7 @@
 <script setup>
 import Container from './Container.vue'
 import { Modal, visitModal } from '@inertiaui/modal-vue'
+import { Link } from '@inertiajs/vue3'
 
 function visitEdit() {
     visitModal('/users/1/edit', {
@@ -16,20 +17,30 @@ function visitEdit() {
 
 <template>
     <Container>
-        <div class="flex justify-between">
+        <div class="">
             <h2 class="text-lg font-medium text-gray-900">Visit programmatically</h2>
+            <div class="flex flex-col items-start">
+                <button @click="visitModal('#local')" type="button">
+                    Open Local Modal
+                </button>
 
-            <button @click="visitModal('#local')" type="button">
-                Open Local Modal
-            </button>
+                <button @click="visitModal('/data', { method: 'post', data: { message: 'Hi again!' } })" type="button">
+                    Open Route Modal
+                </button>
 
-            <button @click="visitModal('/data', { method: 'post', data: { message: 'Hi again!' } })" type="button">
-                Open Route Modal
-            </button>
+                <button @click="visitEdit" type="button">
+                    Open Route Modal With Navigate
+                </button>
+            </div>
+        </div>
 
-            <button @click="visitEdit" type="button">
-                Open Route Modal With Navigate
-            </button>
+        <div class="mt-8">
+            <h2 class="text-lg font-medium text-gray-900">Other stuff</h2>
+            <div class="flex flex-col items-start">
+                <Link href="/conditionally-redirect?redirect=1" dusk="conditional-redirect">
+                    Open page that redirects to modal
+                </Link>
+            </div>
         </div>
     </Container>